fenced-frame/notify-event-success.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/utils.js"></script>
<title>Test fenced frame notifyEvent() succeeds on click</title>

<body>
  <script>
    promise_test(async (t) => {
      const frame_context = await attachFencedFrameContext(
                  {generator_api: 'fledge'});

      var triggered = false;

      // Add the fenced event handler to the HTMLFencedFrameElement.
      frame_context.element.addEventListener('fencedtreeclick', () => {
        triggered = true;
      });

      // Add a click handler to the fenced frame's content, which will trigger
      // the fenced handler registered above on the element.
      await frame_context.execute(() => {
        document.body.addEventListener('click', (e) => {
          window.fence.notifyEvent(e);
        });
      });

      // Move onto the frame and click it.
      var actions = new test_driver.Actions()
      await actions.pointerMove(0, 0, {origin: frame_context.element})
            .pointerDown()
            .pointerUp()
            .send();

      assert_true(triggered);

    }, "Test that a fenced frame notifyEvent() succeeds on click");
  </script>
</body>
